backend/requirements.txt
fastapi[all]==0.95.2
uvicorn==0.22.0
pydantic==1.10.12
pytest==7.4.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.0.0
python-dotenv==1.0.0
